pullup SGComponent
